% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{build_feature_lead_time}
\alias{build_feature_lead_time}
\title{Build a Feature's Lead Time}
\usage{
build_feature_lead_time(dev, pr, build, release)
}
\arguments{
\item{dev}{numeric vector of length n}

\item{pr}{numeric vector of length n}

\item{build}{numeric vector of length n}

\item{release}{numeric vector of length n}
}
\value{
list of length 4 with dev, pr, build, and release times
}
\description{
The most granular unit of software delivery speed is the lead time of a single pull request. There are four
component measurements of a PR's lead time, the sum of which is the aggregate lead time for that PR. The four
components are:
  1. Development -- the length of time of development (i.e. discovery work as well as programming)
  2. Pull Request - the length of time between opening of the pull request and merging it into the master branch
  3. Build - the length of time alloted to automated building of a deployable artifact
  4. Release -- the length of time between a feature being ready for development and it's availability in every
     environment
}
