% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot a single distribution}
\usage{
plot_dist(t, geom, ...)
}
\arguments{
\item{t}{time measurements}

\item{geom}{Geom to use}

\item{...}{Args to pass to geom function}
}
\value{

}
\description{
Plot a single distribution
}
