% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis.R
\name{plot_feature_distribution}
\alias{plot_feature_distribution}
\title{Visualize a feature's distributions}
\usage{
plot_feature_distribution(features)
}
\arguments{
\item{feature}{list of 4 feature components}
}
\value{

}
\description{
Visualize a feature's distributions
}
