% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{rgen}
\alias{rgen}
\title{Closure helper for random generators}
\usage{
rgen(n, fnc, ...)
}
\arguments{
\item{n}{number of samples}

\item{fnc}{random generator to use}

\item{...}{passed to generator function}
}
\value{

}
\description{
Closure helper for random generators
}
\examples{
# Calling .rnorm() is equivalent to calling rnorm(n, mean = 3, sd = 10)
.rnorm <- rgen(10, rnorm, mean = 3, sd = 10)
}
