% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seconds}
\alias{seconds}
\title{Convert a time unit to seconds}
\usage{
seconds(t, unit = NULL)
}
\arguments{
\item{t}{time magnitude}

\item{unit}{time unit}
}
\value{

}
\description{
Convert a time unit to seconds
}
