% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis.R
\name{theme_oracle}
\alias{theme_oracle}
\title{Standard ggplot theme for the agile oracle package. It is thin wrapper around ggplot::theme_minimal().}
\usage{
theme_oracle()
}
\value{

}
\description{
Standard ggplot theme for the agile oracle package. It is thin wrapper around ggplot::theme_minimal().
}
\examples{
ggplot2::ggplot(data = mtcars, mapping = ggplot2::aes(x = disp, y = mpg)) +
  ggplot2::geom_point() +
  theme_oracle() +
  ggplot2::labs(title = "My plot")
}
